requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.34.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scikit-surprise==1.1.3
python-dotenv==1.0.0
sagemaker==2.190.0

